fix(ai): wire locale into text-field prompts and surface keywords/tags
Recipe/ingredient text-field prompts (description, category, cuisine,
summary, culinaryUse, etc.) ignored the request locale, so the model
defaulted to English for short abstract fields even when long fields
echoed the input language. Threaded locale through the prompt and
emit an explicit "Write the output in {LanguageName}" rule, sharing a
localeToLanguageName helper across all three contracts.

Recipe refresh runner added "keywords" to targetFields server-side but
filtered improvements on the client's missingFields list, so the
keywords suggestion was discarded before reaching the client. The
"tags" field config was wired but never requested at all. Both now go
through a unified IMPROVEMENT_FIELDS set so any non-side-effect field
that produced a suggestion lands in aiSuggestions.improvements.

Removed the dead top-level aiSuggestions.tags placeholder.
