+++ apps/website/src/contracts/recipeContract.ts
@@ -1,6 +1,7 @@
 import { z } from "zod";
 import { recipeSchema } from "recipe-ingestion";
 import type { AiContract, FieldConfig } from "@pixelmord/content-ai-refine";
+import { localeToLanguageName } from "./locale-language.ts";
 
 type RecipeSchema = typeof recipeSchema;
 
@@ -15,23 +16,6 @@ export interface RecipeRefineContext {
     recipeIngredient?: string[];
   }>;
   locale?: string;
-}
-
-// Maps BCP-47 locale codes to full language names for stronger model instructions.
-// The model reliably respects "German" far more than the bare two-letter code "de".
-const LOCALE_LANGUAGE_NAMES: Record<string, string> = {
-  en: "English",
-  de: "German",
-  fr: "French",
-  es: "Spanish",
-  it: "Italian",
-  pt: "Portuguese",
-  nl: "Dutch",
-  pl: "Polish",
-  sv: "Swedish",
-};
-function localeToLanguageName(locale: string): string {
-  return LOCALE_LANGUAGE_NAMES[locale] ?? locale;
 }
 
 // Rule reused across string[] fields so suggestions don't echo back values
@@ -82,8 +66,10 @@ const presets = [
 
 // Text improvement field helper (replaces proposeRecipeImprovements per-field)
 const textFieldConfig = (instruction: string): FieldConfig<RecipeSchema, RecipeRefineContext> => ({
-  systemPrompt: ({ currentData }) => {
+  systemPrompt: ({ currentData, sourceContext }) => {
     const ctx = buildRecipeCtx(currentData);
+    const locale = sourceContext?.locale ?? "en";
+    const languageName = localeToLanguageName(locale);
     return `You are a culinary recipe editor.
 
 Recipe context:
@@ -92,6 +78,7 @@ ${ctx}
 ${instruction}
 
 Rules:
+- Write the output in ${languageName}. Do not use any other language.
 - Be specific and informative
 - For time fields use ISO 8601 duration format (e.g. "PT15M", "PT1H30M")
 - Do NOT suggest image URLs`;
